% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{add_pages_to_query_args}
\alias{add_pages_to_query_args}
\title{Add Pages To Query Arguments}
\usage{
add_pages_to_query_args(query_args, n_pages)
}
\arguments{
\item{query_args}{A named list}

\item{n_pages}{An integer}
}
\description{
Add Pages To Query Arguments
}
