% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{create_result_from_paginated_api_query2}
\alias{create_result_from_paginated_api_query2}
\title{Create Result From Paginated API Query2}
\usage{
create_result_from_paginated_api_query2(
  query_args,
  query_file,
  default_tbl,
  unnest_cols = NULL,
  select_cols = NULL,
  arrange_cols = NULL,
  ...
)
}
\arguments{
\item{query_args}{A named list}

\item{query_file}{A string, that is a path to a text file}

\item{default_tbl}{A tibble}

\item{unnest_cols}{A vector of strings passed to tidyr::unnest}

\item{select_cols}{A vector of strings passed to dplyr::select}

\item{arrange_cols}{A vector fo strings passed to dplyr::arrange}

\item{...}{Arguments passed to perform_api_query}
}
\description{
Create Result From Paginated API Query2
}
