% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{get_nested_column}
\alias{get_nested_column}
\title{Get Nested Column}
\usage{
get_nested_column(tbl, column, subcolumn)
}
\arguments{
\item{tbl}{A tibble}

\item{column}{A string that is the name of a column}

\item{subcolumn}{A string that is the name of a subcolumn in the column}
}
\description{
Get Nested Column
}
