% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{perform_api_query}
\alias{perform_api_query}
\title{Perform API Query}
\usage{
perform_api_query(
  variables,
  query_file,
  query_dir = system.file("queries", package = "iatlasGraphQLClient"),
  api_url = "https://api-staging.cri-iatlas.org/api"
)
}
\arguments{
\item{variables}{A named list}

\item{query_file}{A string, that is a path to a text file}

\item{query_dir}{A string, that is to the directory of the query file}

\item{api_url}{A string, that is the URL of the API server}
}
\description{
Perform API Query
}
