% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_cells_queries.R
\name{query_cell_feature_values}
\alias{query_cell_feature_values}
\title{Query Feature Values}
\usage{
query_cell_feature_values(cohorts = NA, cells = NA, paging = NA, ...)
}
\arguments{
\item{cohorts}{A vector of strings}

\item{cells}{A vector of strings}

\item{paging}{A named list}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
Query Feature Values
}
