% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_cell_stats_queries.R
\name{query_cell_stats}
\alias{query_cell_stats}
\title{Query Cell Stats}
\usage{
query_cell_stats(entrez = NA, paging = NA, ...)
}
\arguments{
\item{entrez}{A vector integers}

\item{paging}{A named list}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
Query Cell Stats
}
