% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_cells_queries.R
\name{query_cells}
\alias{query_cells}
\title{Query Cells}
\usage{
query_cells(cohort = NA, cell = NA, paging = NA, ...)
}
\arguments{
\item{cohort}{A vector of characters}

\item{cell}{A vector of characters}

\item{paging}{A named list}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
Query Cells
}
