% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_cohorts_queries.R
\name{query_cohort_mutations}
\alias{query_cohort_mutations}
\title{Query Cohort Mutations}
\usage{
query_cohort_mutations(
  cohorts = NA,
  datasets = NA,
  tags = NA,
  paging = NA,
  ...
)
}
\arguments{
\item{cohorts}{A vector of strings}

\item{datasets}{A vector of strings}

\item{tags}{A vector of strings}

\item{paging}{A named list}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
Query Cohort Mutations
}
