% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_cohorts_queries.R
\name{query_cohort_samples}
\alias{query_cohort_samples}
\title{Query Cohort Samples}
\usage{
query_cohort_samples(
  cohorts = NA,
  datasets = NA,
  tags = NA,
  clinical = NA,
  paging = NA,
  ...
)
}
\arguments{
\item{cohorts}{A vector of strings}

\item{datasets}{A vector of strings}

\item{tags}{A vector of strings}

\item{clinical}{A vector of strings}

\item{paging}{A named list}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
Query Cohort Samples
}
