% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_queries.R
\name{query_cohort_selector}
\alias{query_cohort_selector}
\title{query_cohort_selector}
\usage{
query_cohort_selector(
  datasets,
  parent_tags,
  tags = NA,
  features = NA,
  feature_classes = NA,
  samples = NA,
  ...
)
}
\arguments{
\item{datasets}{A vector of strings}

\item{parent_tags}{A vector of strings}

\item{tags}{A vector of strings}

\item{features}{A vector of strings}

\item{feature_classes}{A vector of strings}

\item{samples}{A vector of strings}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
query_cohort_selector
}
