% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_cohorts_queries.R
\name{query_cohorts}
\alias{query_cohorts}
\title{Query Cohorts}
\usage{
query_cohorts(cohorts = NA, datasets = NA, tags = NA, paging = NA, ...)
}
\arguments{
\item{cohorts}{A vector of strings}

\item{datasets}{A vector of strings}

\item{tags}{A vector of strings}

\item{paging}{A named list}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
Query Cohorts
}
