% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_colocalization_queries.R
\name{query_colocalizations}
\alias{query_colocalizations}
\title{Query Colocalizations}
\usage{
query_colocalizations(
  datasets = NA,
  coloc_datasets = NA,
  features = NA,
  entrez = NA,
  snps = NA,
  qtl_type = NA,
  ecaviar_pp = NA,
  plot_type = NA,
  paging = NA,
  ...
)
}
\arguments{
\item{datasets}{A vector of strings}

\item{coloc_datasets}{A vector of strings}

\item{features}{A vector of strings}

\item{entrez}{A vector of integers}

\item{snps}{A vector of strings}

\item{qtl_type}{A string}

\item{ecaviar_pp}{A string}

\item{plot_type}{A string}

\item{paging}{A named list}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
Query Colocalizations
}
