% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_copy_number_result_queries.R
\name{query_copy_number_results}
\alias{query_copy_number_results}
\title{Query Copy Number Results}
\usage{
query_copy_number_results(
  datasets = NA,
  tags = NA,
  features = NA,
  entrez = NA,
  direction = NA,
  min_p_value = NA,
  max_p_value = NA,
  min_log10_p_value = NA,
  max_log10_p_value = NA,
  min_mean_normal = NA,
  min_mean_cnv = NA,
  min_t_stat = NA,
  paging = NA,
  ...
)
}
\arguments{
\item{datasets}{A vector of strings}

\item{tags}{A vector of strings}

\item{features}{A vector of strings}

\item{entrez}{A vector of strings}

\item{direction}{A string, either Amp' or 'Del'}

\item{min_p_value}{A double}

\item{max_p_value}{A double}

\item{min_log10_p_value}{A double}

\item{max_log10_p_value}{A double}

\item{min_mean_normal}{A double}

\item{min_mean_cnv}{A double}

\item{min_t_stat}{A double}

\item{paging}{A named list}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
Query Copy Number Results
}
