% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_datasets_queries.R
\name{query_dataset_samples}
\alias{query_dataset_samples}
\title{Query Dataset Samples}
\usage{
query_dataset_samples(
  datasets = NA,
  samples = NA,
  types = NA,
  paging = NA,
  ...
)
}
\arguments{
\item{datasets}{A vector of strings that are names of datasets}

\item{samples}{A vector of strings}

\item{types}{A vector of strings}

\item{paging}{A named list}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
Query Dataset Samples
}
