% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_queries2.R
\name{query_dataset_tags}
\alias{query_dataset_tags}
\title{Query Dataset Tags}
\usage{
query_dataset_tags(dataset, ...)
}
\arguments{
\item{dataset}{A string}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
Query Dataset Tags
}
