% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_queries.R
\name{query_datasets}
\alias{query_datasets}
\title{Query Datasets}
\usage{
query_datasets(datasets = NA, ...)
}
\arguments{
\item{datasets}{A vector of strings that are names of datasets}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
Query Datasets
}
