% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_driver_result_queries.R
\name{query_driver_results}
\alias{query_driver_results}
\title{Query Driver Results}
\usage{
query_driver_results(
  datasets = NA,
  tags = NA,
  features = NA,
  entrez = NA,
  mutation_codes = NA,
  min_p_value = NA,
  max_p_value = NA,
  min_log10_p_value = NA,
  max_log10_p_value = NA,
  min_fold_change = NA,
  min_log10_fold_change = NA,
  min_num_wild_types = NA,
  min_num_mutants = NA,
  paging = NA,
  ...
)
}
\arguments{
\item{datasets}{A vector of strings}

\item{tags}{A vector of strings}

\item{features}{A vector of strings}

\item{entrez}{A vector of strings}

\item{mutation_codes}{A vector of strings}

\item{min_p_value}{A double}

\item{max_p_value}{A double}

\item{min_log10_p_value}{A double}

\item{max_log10_p_value}{A double}

\item{min_fold_change}{A double}

\item{min_log10_fold_change}{A double}

\item{min_num_wild_types}{An integer}

\item{min_num_mutants}{An integer}

\item{paging}{A named list}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
Query Driver Results
}
