% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_edges_queries.R
\name{query_edges}
\alias{query_edges}
\title{Query Edges}
\usage{
query_edges(
  max_score = NA,
  min_score = NA,
  node1 = NA,
  node2 = NA,
  paging = NA,
  ...
)
}
\arguments{
\item{max_score}{A double}

\item{min_score}{A double}

\item{node1}{A vector of strings}

\item{node2}{A vector of strings}

\item{paging}{A named list}

\item{...}{Arguments to create_result_from_paginated_api_query}
}
\description{
Query Edges
}
