% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_queries.R
\name{query_expression_by_genes}
\alias{query_expression_by_genes}
\title{Query Expression By Genes}
\usage{
query_expression_by_genes(gene_types = NA, entrez = NA, samples = NA, ...)
}
\arguments{
\item{gene_types}{A vector of strings}

\item{entrez}{A vector of integers}

\item{samples}{A vector of strings}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
Query Expression By Genes
}
