% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_queries.R
\name{query_features}
\alias{query_features}
\title{Query Features}
\usage{
query_features(
  features = NA,
  datasets = NA,
  parent_tags = NA,
  tags = NA,
  feature_classes = NA,
  samples = NA,
  max_value = NA,
  min_value = NA,
  ...
)
}
\arguments{
\item{features}{A vector of strings}

\item{datasets}{A vector of strings}

\item{parent_tags}{A vector of strings}

\item{tags}{A vector of strings}

\item{feature_classes}{A vector of strings}

\item{samples}{A vector of strings}

\item{max_value}{A numeric}

\item{min_value}{A numeric}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
Query Features
}
