% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_queries.R
\name{query_features_by_class}
\alias{query_features_by_class}
\title{Query Features By Class}
\usage{
query_features_by_class(
  datasets = NA,
  parent_tags = NA,
  features = NA,
  feature_classes = NA,
  ...
)
}
\arguments{
\item{datasets}{A vector of strings}

\item{parent_tags}{A vector of strings}

\item{features}{A vector of strings}

\item{feature_classes}{A vector of strings}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
Query Features By Class
}
