% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_cohorts_queries.R
\name{query_full_cohort}
\alias{query_full_cohort}
\title{Query Cohorts}
\usage{
query_full_cohort(cohort, paging = NA, ...)
}
\arguments{
\item{cohort}{A string}

\item{paging}{A named list}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
Query Cohorts
}
