% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_nodes_queries.R
\name{query_gene_nodes}
\alias{query_gene_nodes}
\title{Query Gene Nodes}
\usage{
query_gene_nodes(
  datasets = NA,
  entrez = NA,
  max_score = NA,
  min_score = NA,
  parent_tags = NA,
  network = NA,
  tags = NA,
  page = NA,
  ...
)
}
\arguments{
\item{datasets}{A vector of strings}

\item{entrez}{A vector of integers}

\item{max_score}{A float}

\item{min_score}{A float}

\item{parent_tags}{A vector of strings}

\item{network}{A vector of strings}

\item{tags}{A vector of strings}

\item{page}{An integer}

\item{...}{Arguments to create_result_from_paginated_api_query}
}
\description{
Query Gene Nodes
}
