% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_gene_type.R
\name{query_gene_types}
\alias{query_gene_types}
\title{Query Gene Types}
\usage{
query_gene_types(gene_types = NA, ...)
}
\arguments{
\item{gene_types}{A vector of strings}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
Query Gene Types
}
