% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_queries.R
\name{query_genes}
\alias{query_genes}
\title{Query Genes}
\usage{
query_genes(gene_types = NA, entrez = NA, samples = NA, ...)
}
\arguments{
\item{gene_types}{A vector of strings}

\item{entrez}{A vector of integers}

\item{samples}{A vector of strings}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
Query Genes
}
