% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_queries.R
\name{query_genes_expression_by_tag}
\alias{query_genes_expression_by_tag}
\title{Query Genes Expression By Tag}
\usage{
query_genes_expression_by_tag(
  datasets,
  parent_tags,
  tags = NA,
  entrez = NA,
  gene_types = NA,
  features = NA,
  feature_classes = NA,
  samples = NA,
  ...
)
}
\arguments{
\item{datasets}{A vector of strings}

\item{parent_tags}{A vector of strings}

\item{tags}{A vector of strings}

\item{entrez}{A vector of integers}

\item{gene_types}{A vector of strings}

\item{features}{A vector of strings}

\item{feature_classes}{A vector of strings}

\item{samples}{A vector of strings}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
Query Genes Expression By Tag
}
