% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_germline_gwas_results_queries.R
\name{query_germline_gwas_results}
\alias{query_germline_gwas_results}
\title{Query Germline Gwas Results}
\usage{
query_germline_gwas_results(
  datasets = NA,
  features = NA,
  snps = NA,
  min_p_value = NA,
  max_p_value = NA,
  paging = NA,
  ...
)
}
\arguments{
\item{datasets}{A vector of strings}

\item{features}{A vector of strings}

\item{snps}{A vector of strings}

\item{min_p_value}{A double}

\item{max_p_value}{A double}

\item{paging}{A named list}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
Query Germline Gwas Results
}
