% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_heritability_result_queries.R
\name{query_heritability_results}
\alias{query_heritability_results}
\title{Query Heritability Results}
\usage{
query_heritability_results(
  datasets = NA,
  features = NA,
  clusters = NA,
  min_p_value = NA,
  max_p_value = NA,
  paging = NA,
  ...
)
}
\arguments{
\item{datasets}{A vector of strings}

\item{features}{A vector of strings}

\item{clusters}{A vector of strings}

\item{min_p_value}{A double}

\item{max_p_value}{A double}

\item{paging}{A named list}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
Query Heritability Results
}
