% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_genes_queries.R
\name{query_immunomodulators}
\alias{query_immunomodulators}
\title{Query Immunomodulators}
\usage{
query_immunomodulators(type = "immunomodulator", entrez = NA, ...)
}
\arguments{
\item{type}{A vector of strings}

\item{entrez}{A vector of integers}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
Query Immunomodulators
}
