% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_mutations_queries.R
\name{query_mutation_statuses}
\alias{query_mutation_statuses}
\title{Query Mutation Statuses}
\usage{
query_mutation_statuses(
  ids = NA,
  entrez = NA,
  codes = NA,
  types = NA,
  cohorts = NA,
  samples = NA,
  status = NA,
  paging = NA,
  ...
)
}
\arguments{
\item{ids}{A vector of integers}

\item{entrez}{A vector of integers}

\item{codes}{A vector of strings}

\item{types}{A vector of strings}

\item{cohorts}{A vector of strings}

\item{samples}{A vector of strings}

\item{status}{A string}

\item{paging}{A named list}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
Query Mutation Statuses
}
