% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_mutations_queries.R
\name{query_mutations}
\alias{query_mutations}
\title{query_mutations}
\usage{
query_mutations(
  datasets = NA,
  ids = NA,
  entrez = NA,
  codes = NA,
  types = NA,
  parent_tags = NA,
  samples = NA,
  status = NA,
  tags = NA,
  ...
)
}
\arguments{
\item{datasets}{A vector of strings}

\item{ids}{A vector of integers}

\item{entrez}{A vector of integers}

\item{codes}{A vector of strings}

\item{types}{A vector of strings}

\item{parent_tags}{A vector of strings}

\item{samples}{A vector of strings}

\item{status}{A string}

\item{tags}{A vector of strings}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
query_mutations
}
