% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_queries2.R
\name{query_mutations_by_samples}
\alias{query_mutations_by_samples}
\title{Query Mutations BY Samples}
\usage{
query_mutations_by_samples(
  entrez = NA,
  mutation_codes = NA,
  mutation_types = NA,
  mutation_ids = NA,
  mutation_status = NA,
  samples = NA,
  page = NA,
  ...
)
}
\arguments{
\item{entrez}{A vector of integers}

\item{mutation_codes}{A vector of strings}

\item{mutation_types}{A vector of strings}

\item{mutation_ids}{A string}

\item{mutation_status}{An integer}

\item{samples}{A vector of strings}

\item{page}{An integer}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
Query Mutations BY Samples
}
