% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_neoantigens_queries.R
\name{query_neoantigens}
\alias{query_neoantigens}
\title{Query Neoantigens}
\usage{
query_neoantigens(pmhcs = NA, entrez = NA, patients = NA, paging = NA, ...)
}
\arguments{
\item{pmhcs}{A vector of strings}

\item{entrez}{A vector of integers}

\item{patients}{A vector of strings}

\item{paging}{A named list}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
Query Neoantigens
}
