% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_nodes_queries.R
\name{query_nodes}
\alias{query_nodes}
\title{Query Nodes}
\usage{
query_nodes(
  datasets = NA,
  features = NA,
  entrez = NA,
  max_score = NA,
  min_score = NA,
  parent_tags = NA,
  network = NA,
  tag1 = NA,
  tag2 = NA,
  n_tags = NA,
  paging = NA,
  ...
)
}
\arguments{
\item{datasets}{A vector of strings}

\item{features}{A vector of strings}

\item{entrez}{A vector of integers}

\item{max_score}{A float}

\item{min_score}{A float}

\item{parent_tags}{A vector of strings}

\item{network}{A vector of strings}

\item{tag1}{A vector of strings}

\item{tag2}{A vector of strings}

\item{n_tags}{An integer, either 1 or 2}

\item{paging}{A named list}

\item{...}{Arguments to create_result_from_paginated_api_query}
}
\description{
Query Nodes
}
