% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_queries2.R
\name{query_patient_slides}
\alias{query_patient_slides}
\title{Query Patient Slides}
\usage{
query_patient_slides(
  patients = NA,
  datasets = NA,
  ethnicities = NA,
  genders = NA,
  races = NA,
  samples = NA,
  slides = NA,
  max_age_at_diagnosis = NA,
  max_height = NA,
  max_weight = NA,
  min_age_at_diagnosis = NA,
  min_height = NA,
  min_weight = NA,
  ...
)
}
\arguments{
\item{patients}{A list of strings}

\item{datasets}{A vector of strings}

\item{ethnicities}{A vector of strings}

\item{genders}{A vector of strings}

\item{races}{A vector of strings}

\item{samples}{A vector of strings}

\item{slides}{A vector of strings}

\item{max_age_at_diagnosis}{An integer}

\item{max_height}{A float}

\item{max_weight}{A float}

\item{min_age_at_diagnosis}{An integer}

\item{min_height}{A float}

\item{min_weight}{A float}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
Query Patient Slides
}
