% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_queries.R
\name{query_patients}
\alias{query_patients}
\title{Query Patients}
\usage{
query_patients(patients, ...)
}
\arguments{
\item{patients}{A list of strings}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
Query Patients
}
