% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_queries.R
\name{query_sample_patients}
\alias{query_sample_patients}
\title{Query Sample Patients}
\usage{
query_sample_patients(samples = NA, patients = NA, ...)
}
\arguments{
\item{samples}{A list of strings}

\item{patients}{A list of strings}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
Query Sample Patients
}
