% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_queries.R
\name{query_samples}
\alias{query_samples}
\title{Query Samples}
\usage{
query_samples(samples = NA, patients = NA, ...)
}
\arguments{
\item{samples}{A list of strings}

\item{patients}{A list of strings}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
Query Samples
}
