% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_samples_by_tags_queries.R
\name{query_samples_by_tag2}
\alias{query_samples_by_tag2}
\title{Query Samples By Tag 2}
\usage{
query_samples_by_tag2(
  datasets = NA,
  parent_tags = NA,
  tags = NA,
  features = NA,
  feature_classes = NA,
  samples = NA,
  patients = NA,
  ...
)
}
\arguments{
\item{datasets}{A vector of strings}

\item{parent_tags}{A vector of strings}

\item{tags}{A vector of strings}

\item{features}{A vector of strings}

\item{feature_classes}{A vector of strings}

\item{samples}{A vector of strings}

\item{patients}{A vector of strings}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
Query Samples By Tag 2
}
