% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_cells_queries.R
\name{query_single_cell_feature}
\alias{query_single_cell_feature}
\title{Query SingleCellFeature}
\usage{
query_single_cell_feature(
  feature = NA,
  cohort = NA,
  cell = NA,
  paging = NA,
  ...
)
}
\arguments{
\item{feature}{A vector of characters}

\item{cohort}{A vector of characters}

\item{cell}{A vector of characters}

\item{paging}{A named list}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
Query SingleCellFeature
}
