% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_cells_queries.R
\name{query_single_cell_seq}
\alias{query_single_cell_seq}
\title{Query SingleCellSeq}
\usage{
query_single_cell_seq(entrez = NA, cohort = NA, cell = NA, paging = NA, ...)
}
\arguments{
\item{entrez}{A vector integers}

\item{cohort}{A vector of characters}

\item{cell}{A vector of characters}

\item{paging}{A named list}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
Query SingleCellSeq
}
