% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_features_queries.R
\name{query_single_cell_seq_feature_values}
\alias{query_single_cell_seq_feature_values}
\title{Query Single Cell Seq Feature Values}
\usage{
query_single_cell_seq_feature_values(
  cohorts = NA,
  features = NA,
  paging = NA,
  ...
)
}
\arguments{
\item{cohorts}{A vector of strings}

\item{features}{A vector of strings}

\item{paging}{A named list}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
Query Single Cell Seq Feature Values
}
