% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_slides_queries.R
\name{query_slides}
\alias{query_slides}
\title{query_slides}
\usage{
query_slides(slides = NA, paging = NA, ...)
}
\arguments{
\item{slides}{A vector of strings}

\item{paging}{A named list}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
query_slides
}
