% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_snps_queries.R
\name{query_snps}
\alias{query_snps}
\title{Query Snps}
\usage{
query_snps(
  name = NA,
  rsid = NA,
  chr = NA,
  min_bp = NA,
  max_bp = NA,
  paging = NA,
  ...
)
}
\arguments{
\item{name}{A vector of strings}

\item{rsid}{A vector of strings}

\item{chr}{A vector of strings}

\item{min_bp}{An integer}

\item{max_bp}{An}

\item{paging}{A named list}

\item{...}{Arguments to create_result_from_paginated_api_query}
}
\description{
Query Snps
}
