% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_queries.R
\name{query_tag_samples}
\alias{query_tag_samples}
\title{Query Tag Samples}
\usage{
query_tag_samples(
  datasets = NA,
  parent_tags = NA,
  tags = NA,
  features = NA,
  feature_classes = NA,
  sample_names = NA,
  patients = NA,
  ...
)
}
\arguments{
\item{datasets}{A vector of strings}

\item{parent_tags}{A vector of strings}

\item{tags}{A vector of strings}

\item{features}{A vector of strings}

\item{feature_classes}{A vector of strings}

\item{sample_names}{A vector of strings}

\item{patients}{A vector of strings}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
Query Tag Samples
}
