% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_tags_queries.R
\name{query_tag_samples2}
\alias{query_tag_samples2}
\title{Tag Samples2}
\usage{
query_tag_samples2(
  datasets,
  parent_tags,
  tags = NA,
  features = NA,
  feature_classes = NA,
  samples = NA,
  ...
)
}
\arguments{
\item{datasets}{A vector of strings}

\item{parent_tags}{A vector of strings}

\item{tags}{A vector of strings}

\item{features}{A vector of strings}

\item{feature_classes}{A vector of strings}

\item{samples}{A vector of strings}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
Tag Samples2
}
