% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_tags_queries.R
\name{query_tag_samples_parents}
\alias{query_tag_samples_parents}
\title{Query Tag Samples Parents}
\usage{
query_tag_samples_parents(
  cohorts = NA_character_,
  samples = NA_character_,
  datasets = NA_character_,
  parent_tags = NA_character_,
  tags = NA_character_,
  type = NA_character_,
  paging = NA,
  ...
)
}
\arguments{
\item{cohorts}{a vector of strings}

\item{samples}{A vector of strings}

\item{datasets}{A vector of strings}

\item{parent_tags}{A vector of strings}

\item{tags}{A vector of strings}

\item{type}{A vector of strings}

\item{paging}{A named list}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
Query Tag Samples Parents
}
