% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{separate_combined_column}
\alias{separate_combined_column}
\title{Separate Combined Column}
\usage{
separate_combined_column(tbl, column)
}
\arguments{
\item{tbl}{A tibble}

\item{column}{A string that is the name of a column}
}
\description{
Separate Combined Column
}
