% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{format_query_result}
\alias{format_query_result}
\title{Format Query Result}
\usage{
format_query_result(tbl, select_cols, arrange_cols)
}
\arguments{
\item{tbl}{A Tibble}

\item{select_cols}{A vector of strings passed to dplyr::select}

\item{arrange_cols}{A vector fo strings passed to dplyr::arrange}
}
\description{
Format Query Result
}
