% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_queries.R
\name{query_dataset_samples}
\alias{query_dataset_samples}
\title{Query Dataset Samples}
\usage{
query_dataset_samples(datasets, ...)
}
\arguments{
\item{datasets}{A vector of strings that are names of datasets}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
Query Dataset Samples
}
