% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_features_queries.R
\name{query_feature_values}
\alias{query_feature_values}
\title{Query Feature Values}
\usage{
query_feature_values(
  datasets = NA,
  parent_tags = NA,
  tags = NA,
  features = NA,
  feature_classes = NA,
  samples = NA,
  max_value = NA,
  min_value = NA,
  paging = NA,
  ...
)
}
\arguments{
\item{datasets}{A vector of strings}

\item{parent_tags}{A vector of strings}

\item{tags}{A vector of strings}

\item{features}{A vector of strings}

\item{feature_classes}{A vector of strings}

\item{samples}{A vector of strings}

\item{max_value}{A numeric}

\item{min_value}{A numeric}

\item{paging}{A named list}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
Query Feature Values
}
