% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_features_queries.R
\name{query_features}
\alias{query_features}
\title{Query Features}
\usage{
query_features(
  cohorts = NA,
  samples = NA,
  features = NA,
  feature_classes = NA,
  max_value = NA,
  min_value = NA,
  paging = NA,
  ...
)
}
\arguments{
\item{cohorts}{A vector of strings}

\item{samples}{A vector of strings}

\item{features}{A vector of strings}

\item{feature_classes}{A vector of strings}

\item{max_value}{A numeric}

\item{min_value}{A numeric}

\item{paging}{A named list}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
Query Features
}
