% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_genes_queries.R
\name{query_io_targets}
\alias{query_io_targets}
\title{Query IO Targets}
\usage{
query_io_targets(gene_types = "io_target", entrez = NA, paging = NA, ...)
}
\arguments{
\item{gene_types}{A vector of strings}

\item{entrez}{A vector of integers}

\item{paging}{A named list}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
Query IO Targets
}
