% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_queries2.R
\name{query_mutations}
\alias{query_mutations}
\title{query_mutations}
\usage{
query_mutations(ids = NA, entrez = NA, codes = NA, types = NA, ...)
}
\arguments{
\item{ids}{A vector of integers}

\item{entrez}{A vector of integers}

\item{codes}{A vector of strings}

\item{types}{A vector of strings}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
query_mutations
}
