% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_samples_queries.R
\name{query_sample_patients}
\alias{query_sample_patients}
\title{Query Sample Patients}
\usage{
query_sample_patients(
  samples = NA,
  patients = NA,
  ethnicities = NA,
  genders = NA,
  races = NA,
  max_age_at_diagnosis = NA,
  max_height = NA,
  max_weight = NA,
  min_age_at_diagnosis = NA,
  min_height = NA,
  min_weight = NA,
  paging = NA,
  ...
)
}
\arguments{
\item{samples}{A vector of strings}

\item{patients}{A list of strings}

\item{ethnicities}{A vector of strings}

\item{genders}{A vector of strings}

\item{races}{A vector of strings}

\item{max_age_at_diagnosis}{An integer}

\item{max_height}{A float}

\item{max_weight}{A float}

\item{min_age_at_diagnosis}{An integer}

\item{min_height}{A float}

\item{min_weight}{A float}

\item{paging}{A named list}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
Query Sample Patients
}
