% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_queries2.R
\name{query_samples_by_mutation_status}
\alias{query_samples_by_mutation_status}
\title{query_samples_by_mutation_status}
\usage{
query_samples_by_mutation_status(ids = NA, status = NA, samples = NA, ...)
}
\arguments{
\item{ids}{A vector of integers}

\item{status}{A string, either "Wt" or "Mut"}

\item{samples}{A vector of strings}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
query_samples_by_mutation_status
}
