% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_genes_queries.R
\name{query_single_cell_seq}
\alias{query_single_cell_seq}
\title{Query Single Cell Seq}
\usage{
query_single_cell_seq(cohorts = NA, entrez = NA, paging = NA, ...)
}
\arguments{
\item{cohorts}{A vector of strings}

\item{entrez}{A vector of integers}

\item{paging}{A named list}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
Query Single Cell Seq
}
