% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_tags_queries.R
\name{query_tag_publications}
\alias{query_tag_publications}
\title{Query Tag Publications}
\usage{
query_tag_publications(
  cohorts = NA,
  samples = NA,
  datasets = NA,
  parent_tags = NA,
  tags = NA,
  type = NA,
  paging = NA,
  ...
)
}
\arguments{
\item{cohorts}{a vector of strings}

\item{samples}{A vector of strings}

\item{datasets}{A vector of strings}

\item{parent_tags}{A vector of strings}

\item{tags}{A vector of strings}

\item{type}{A vector of strings}

\item{paging}{A named list}

\item{...}{Arguments to create_result_from_api_query}
}
\description{
Query Tag Publications
}
